Refactored code into a url-helper
