+++ src/shared/util/url-helper.js
@@ -2,3 +2,7 @@
 export function eventHref(event) {
   return `/about-us/events/${event.datetime.year}/${event.datetime.month}/${event.datetime.date}/${event.slug}`;
 }
+
+export function newsItemHref(news) {
+  return `/about-us/news/${news.datetime.year}/${news.datetime.month}/${news.datetime.date}/${news.slug}`;
+}
